refactor(core): oidc provider event listener
