+++ packages/core/src/oidc/init.ts
@@ -12,7 +12,7 @@ import { isOriginAllowed, validateCustomClientMetadata } from '@/oidc/utils';
 import { findResourceByIndicator } from '@/queries/resource';
 import { findUserById } from '@/queries/user';
 import { routes } from '@/routes/consts';
-import { initTokenExchangeSuccessLogger } from '@/utils/logger';
+import { grantSuccessListener } from '@/utils/oidc-provider-event-listener';
 
 export default async function initOidc(app: Koa): Promise<Provider> {
   const { issuer, privateKey, defaultIdTokenTtl, defaultRefreshTokenTtl } = envSet.values.oidc;
@@ -118,7 +118,7 @@ export default async function initOidc(app: Koa): Promise<Provider> {
     },
   });
 
-  oidc.on('grant.success', initTokenExchangeSuccessLogger());
+  oidc.on('grant.success', grantSuccessListener);
 
   app.use(mount('/oidc', oidc.app));
